[TkTrackingRegions] Move public headers to interface directory
