+++ RecoTracker/TkTrackingRegions/interface/GlobalTrackingRegionProducerFromBeamSpot.h
